requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
django-cors-headers
